Make 'inWhichBin' an int

+++ src/pyasl/asl/binning.py
@@ -156,7 +156,7 @@ def binningx0dt(x, y, yerr=None, x0=None, dt=None, nbins=None, reduceBy=None, re
     dt = (np.max(x) - x0)/float(nbins)
   # Start calculation
   # In which bin do the individual data points belong?
-  inWhichBin = np.floor(((x-x0)/dt))
+  inWhichBin = np.floor(((x-x0)/dt)).astype(np.int)
   # Lonely last bin correction
   # Brings the last data point into the last valid bin
   # instead of creating a new bin with that data point\
